+++ src/attribute.ml
@@ -2,7 +2,7 @@
 open Monad.Notations
 
 type t =
-  | Axiom
+  | AxiomWithReason
   | ForceGadt
   | Implicit of string
   | MatchGadt
@@ -44,9 +44,14 @@ let of_attributes (attributes : Typedtree.attributes) : t list Monad.t =
     let id = attr_name.Asttypes.txt in
     match id with
     | "coq_axiom" ->
+      raise
+        None
+        Unexpected
+        "Depreacated attribute. Use @coq_axiom_with_reason instead."
+    | "coq_axiom_with_reason" ->
       let error_message = "Give a reason for this axiom." in
       let* _ = of_payload_string error_message id attr_payload in
-      return (Some Axiom)
+      return (Some AxiomWithReason)
     | "coq_force_gadt" -> return (Some ForceGadt)
     | "coq_implicit" ->
       let error_message =
@@ -65,9 +70,9 @@ let of_attributes (attributes : Typedtree.attributes) : t list Monad.t =
     | _ -> return None)
   )
 
-let has_axiom (attributes : t list) : bool =
+let has_axiom_with_reason (attributes : t list) : bool =
   attributes |> List.exists (function
-    | Axiom -> true
+    | AxiomWithReason -> true
     | _ -> false
   )