+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+* Rename the attribute `@coq_axiom` as `@coq_axiom_with_reason`.
 * Add the attribute `@coq_precise_signature` to help to distinguish between ambiguous signatures.
 * Remove the generation of `Import` in Coq and expand the names regardless of the `open` commands, in order to be compatible with local opens.
 * Add support for anonymous sub-signatures.